+++ pandas/core/series.py
@@ -62,7 +62,8 @@ _shared_doc_kwargs = dict(
     axes='index', klass='Series', axes_single_arg="{0, 'index'}",
     inplace="""inplace : boolean, default False
         If True, performs operation inplace and returns None.""",
-    duplicated='Series')
+    duplicated='Series',
+    optional_by='')
 
 
 def _coerce_method(converter):
